Added sync to master merge
Currently, the sync job is executed only by the schedule and not when something is merged. This means that changes to the script are not verified right away, but only after some time. Hence, I changed it so that the script is also executed whenever there is a merge to master.